pyproject.toml
[tool.poetry]
name = "test_project"
version = "0.1.0"
description = ""
authors = ["khazidhea <azamat@galimzhanov.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-django = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
